refactor: move commonly used COSE functions to utility module
