+++ src/token/cose/crypto_impl/rustcrypto/mac/mod.rs
@@ -8,11 +8,12 @@
  *
  * SPDX-License-Identifier: MIT OR Apache-2.0
  */
-use coset::iana::Algorithm;
 
-use crate::error::CoseCipherError;
+use crate::token::cose::crypto_impl::rustcrypto::CoseCipherError;
 use crate::token::cose::crypto_impl::rustcrypto::RustCryptoContext;
-use crate::token::cose::{CoseSymmetricKey, MacCryptoBackend};
+use crate::token::cose::CoseSymmetricKey;
+use crate::token::cose::MacCryptoBackend;
+use coset::iana;
 use rand::{CryptoRng, RngCore};
 
 #[cfg(feature = "rustcrypto-hmac")]
@@ -22,7 +23,7 @@ impl<RNG: RngCore + CryptoRng> MacCryptoBackend for RustCryptoContext<RNG> {
     #[cfg(feature = "rustcrypto-hmac")]
     fn compute_hmac(
         &mut self,
-        algorithm: Algorithm,
+        algorithm: iana::Algorithm,
         key: CoseSymmetricKey<'_, Self::Error>,
         payload: &[u8],
     ) -> Result<Vec<u8>, CoseCipherError<Self::Error>> {
@@ -32,7 +33,7 @@ impl<RNG: RngCore + CryptoRng> MacCryptoBackend for RustCryptoContext<RNG> {
     #[cfg(feature = "rustcrypto-hmac")]
     fn verify_hmac(
         &mut self,
-        algorithm: Algorithm,
+        algorithm: iana::Algorithm,
         key: CoseSymmetricKey<'_, Self::Error>,
         tag: &[u8],
         payload: &[u8],
