+++ src/token/cose/key.rs
@@ -14,10 +14,9 @@ use core::borrow::Borrow;
 use core::fmt::Display;
 use core::marker::PhantomData;
 use coset::iana::EnumI64;
-use coset::{iana, Algorithm, AsCborValue, CoseKey, KeyType, Label, RegisteredLabelWithPrivate};
+use coset::{iana, AsCborValue, CoseKey, KeyType, Label, RegisteredLabelWithPrivate};
 
 use crate::error::CoseCipherError;
-use crate::token::cose::CryptoBackend;
 
 /// Finds a key parameter by its label.
 fn find_param_by_label<'a>(label: &Label, param_vec: &'a [(Label, Value)]) -> Option<&'a Value> {
@@ -474,187 +473,4 @@ impl<T: KeyProvider> KeyProvider for &T {
     fn lookup_key(&self, key_id: Option<&[u8]>) -> impl Iterator<Item = impl Borrow<CoseKey>> {
         (*self).lookup_key(key_id)
     }
-}
-
-/// Determines the key size that a key for the given `algorithm` should have.
-fn symmetric_key_size<BE: Display>(
-    algorithm: iana::Algorithm,
-) -> Result<usize, CoseCipherError<BE>> {
-    match algorithm {
-        iana::Algorithm::A128GCM
-        | iana::Algorithm::AES_CCM_16_64_128
-        | iana::Algorithm::AES_CCM_64_64_128
-        | iana::Algorithm::AES_CCM_16_128_128
-        | iana::Algorithm::AES_CCM_64_128_128
-        | iana::Algorithm::A128KW => Ok(16),
-        iana::Algorithm::A192GCM | iana::Algorithm::A192KW => Ok(24),
-        iana::Algorithm::A256GCM
-        | iana::Algorithm::AES_CCM_16_64_256
-        | iana::Algorithm::AES_CCM_64_64_256
-        | iana::Algorithm::AES_CCM_16_128_256
-        | iana::Algorithm::AES_CCM_64_128_256
-        | iana::Algorithm::A256KW => Ok(32),
-        _ => Err(CoseCipherError::UnsupportedAlgorithm(Algorithm::Assigned(
-            algorithm,
-        ))),
-    }
-}
-
-/// Attempts to parse the given `parsed_key` as an AES symmetric key.
-///
-/// Performs the checks required for symmetric keys suitable for AES according to
-/// [RFC 9053, Section 4](https://datatracker.ietf.org/doc/html/rfc9053#section-4).
-pub(crate) fn ensure_valid_aes_key<BE: Display>(
-    algorithm: iana::Algorithm,
-    parsed_key: CoseParsedKey<BE>,
-) -> Result<CoseSymmetricKey<BE>, CoseCipherError<BE>> {
-    // Checks according to RFC 9053, Section 4.1 and 4.2.
-
-    // Key type must be symmetric.
-    let symm_key = if let CoseParsedKey::Symmetric(symm_key) = parsed_key {
-        symm_key
-    } else {
-        return Err(CoseCipherError::KeyTypeAlgorithmMismatch(
-            parsed_key.as_ref().kty.clone(),
-            Algorithm::Assigned(algorithm),
-        ));
-    };
-
-    // Algorithm in key must match algorithm to use.
-    if let Some(key_alg) = &symm_key.as_ref().alg {
-        if key_alg != &Algorithm::Assigned(algorithm) {
-            return Err(CoseCipherError::KeyAlgorithmMismatch(
-                key_alg.clone(),
-                Algorithm::Assigned(algorithm),
-            ));
-        }
-    }
-
-    // For algorithms that we know, check the key length (would lead to a cipher error later on).
-    let key_len = symmetric_key_size(algorithm)?;
-    if symm_key.k.len() != key_len {
-        return Err(CoseCipherError::InvalidKeyParam(
-            KeyParam::Symmetric(iana::SymmetricKeyParameter::K),
-            Value::Bytes(symm_key.k.to_vec()),
-        ));
-    }
-
-    Ok(symm_key)
-}
-
-/// Generates a random content encryption key for the given `algorithm` using the given `backend`.
-pub(crate) fn generate_cek_for_alg<B: CryptoBackend>(
-    backend: &mut B,
-    algorithm: iana::Algorithm,
-) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
-    let key_len = symmetric_key_size(algorithm)?;
-    let mut key = vec![0u8; key_len];
-    backend.generate_rand(key.as_mut_slice())?;
-    Ok(key)
-}
-
-/// Attempts to parse the given `parsed_key` as an ECDSA key.
-///
-/// Performs the checks required for ECDSA keys according to
-/// [RFC 9053, Section 2.1](https://datatracker.ietf.org/doc/html/rfc9053#section-2.1) and/or
-/// [RFC 8812, Section 3.2](https://datatracker.ietf.org/doc/html/rfc8812#section-3.2).
-pub(crate) fn ensure_valid_ecdsa_key<BE: Display>(
-    algorithm: iana::Algorithm,
-    parsed_key: CoseParsedKey<BE>,
-    key_should_be_private: bool,
-) -> Result<CoseEc2Key<BE>, CoseCipherError<BE>> {
-    // Checks according to RFC 9053, Section 2.1 or RFC 8812, Section 3.2.
-
-    // Key type must be EC2
-    let ec2_key = if let CoseParsedKey::Ec2(ec2_key) = parsed_key {
-        ec2_key
-    } else {
-        return Err(CoseCipherError::KeyTypeAlgorithmMismatch(
-            parsed_key.as_ref().kty.clone(),
-            Algorithm::Assigned(algorithm),
-        ));
-    };
-
-    // If algorithm in key is set, it must match our algorithm
-    if let Some(key_alg) = &ec2_key.as_ref().alg {
-        if key_alg != &Algorithm::Assigned(algorithm) {
-            return Err(CoseCipherError::KeyAlgorithmMismatch(
-                key_alg.clone(),
-                Algorithm::Assigned(algorithm),
-            ));
-        }
-    }
-
-    // Key must contain private key information to perform signature, and either D or X and Y to
-    // verify a signature.
-    if key_should_be_private && ec2_key.d.is_none() {
-        return Err(CoseCipherError::MissingKeyParam(vec![
-            iana::Ec2KeyParameter::D.into(),
-        ]));
-    } else if !key_should_be_private && ec2_key.d.is_none() {
-        if ec2_key.x.is_none() {
-            return Err(CoseCipherError::MissingKeyParam(vec![
-                iana::Ec2KeyParameter::X.into(),
-                iana::Ec2KeyParameter::D.into(),
-            ]));
-        }
-        if ec2_key.y.is_none() {
-            return Err(CoseCipherError::MissingKeyParam(vec![
-                iana::Ec2KeyParameter::Y.into(),
-                iana::Ec2KeyParameter::D.into(),
-            ]));
-        }
-    }
-
-    Ok(ec2_key)
-}
-
-/// Attempts to parse the given `parsed_key` as an HMAC key.
-///
-/// Performs the checks required for symmetric keys suitable for HMAC according to
-/// [RFC 9053, Section 3.1](https://datatracker.ietf.org/doc/html/rfc9053#section-3.1).
-pub(crate) fn ensure_valid_hmac_key<BE: Display>(
-    algorithm: iana::Algorithm,
-    parsed_key: CoseParsedKey<BE>,
-) -> Result<CoseSymmetricKey<BE>, CoseCipherError<BE>> {
-    // Checks according to RFC 9053, Section 3.1.
-
-    // Key type must be symmetric.
-    let symm_key = if let CoseParsedKey::Symmetric(symm_key) = parsed_key {
-        symm_key
-    } else {
-        return Err(CoseCipherError::KeyTypeAlgorithmMismatch(
-            parsed_key.as_ref().kty.clone(),
-            Algorithm::Assigned(algorithm),
-        ));
-    };
-
-    // Algorithm in key must match algorithm to use.
-    if let Some(key_alg) = &symm_key.as_ref().alg {
-        if key_alg != &Algorithm::Assigned(algorithm) {
-            return Err(CoseCipherError::KeyAlgorithmMismatch(
-                key_alg.clone(),
-                Algorithm::Assigned(algorithm),
-            ));
-        }
-    }
-
-    // For algorithms that we know, check the key length (would lead to a cipher error later on).
-    let key_len = match algorithm {
-        iana::Algorithm::HMAC_256_256 | iana::Algorithm::HMAC_256_64 => Some(32),
-        iana::Algorithm::HMAC_384_384 => Some(48),
-        iana::Algorithm::HMAC_512_512 => Some(64),
-        _ => None,
-    };
-
-    if let Some(key_len) = key_len {
-        if symm_key.k.len() != key_len {
-            return Err(CoseCipherError::InvalidKeyParam(
-                KeyParam::Symmetric(iana::SymmetricKeyParameter::K),
-                Value::Bytes(symm_key.k.to_vec()),
-            ));
-        }
-    }
-
-    Ok(symm_key)
 }
