+++ src/token/cose/mod.rs
@@ -214,19 +214,22 @@ mod signed;
 mod maced;
 mod recipient;
 
+mod header;
+
 /// AAD providers and operations for those.
 pub mod aad;
 
 pub use aad::AadProvider;
 pub use encrypted::*;
+pub use header::*;
 pub use key::*;
 pub use maced::*;
 pub use recipient::*;
 pub use signed::*;
-pub use util::*;
 
 #[cfg(all(test, feature = "std"))]
 pub(crate) mod test_helper;
+/// Utility functions for the cose module itself and for cryptographic backend implementors.
 pub mod util;
 
 /// Trait for implementations of cryptographic functions that can be used for COSE structures.
