+++ src/token/cose/util/header.rs
@@ -15,33 +15,10 @@ use core::fmt::Display;
 
 use alloc::borrow::Borrow;
 use coset::iana::EnumI64;
-use coset::{iana, Algorithm, CoseKey, Header, HeaderBuilder, KeyOperation, Label};
+use coset::{iana, Algorithm, CoseKey, Header, KeyOperation, Label};
 
 use crate::error::CoseCipherError;
 use crate::token::cose::key::KeyProvider;
-use crate::token::cose::{aes_algorithm_iv_len, CryptoBackend, EncryptCryptoBackend};
-
-/// A header parameter that can be used in a COSE header.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum HeaderParam {
-    /// Generic header parameter applicable to all algorithms.
-    Generic(iana::HeaderParameter),
-    /// Header parameter that is specific for a set of algorithms.
-    Algorithm(iana::HeaderAlgorithmParameter),
-}
-
-impl From<iana::HeaderParameter> for HeaderParam {
-    fn from(value: iana::HeaderParameter) -> Self {
-        HeaderParam::Generic(value)
-    }
-}
-
-impl From<iana::HeaderAlgorithmParameter> for HeaderParam {
-    fn from(value: iana::HeaderAlgorithmParameter) -> Self {
-        HeaderParam::Algorithm(value)
-    }
-}
-
 /// Returns the set of header parameters that are set in the given `header_bucket`.
 fn create_header_parameter_set(header_bucket: &Header) -> BTreeSet<Label> {
     let mut header_bucket_fields = BTreeSet::new();
@@ -174,79 +151,4 @@ pub(crate) fn determine_key_candidates<'a, CKP: KeyProvider, CE: Display>(
         }
         Ok((k_borrow.to_owned(), chosen_alg))
     })
-}
-
-/// Extensions to the [`HeaderBuilder`] type that enable usage of cryptographic backends.
-pub trait HeaderBuilderExt: Sized {
-    /// Generate an initialization vector for the given `algorithm` using the given
-    /// cryptographic `backend`.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the `algorithm` is unsupported/unknown or the cryptographic backend
-    /// returns an error.
-    fn gen_iv<B: EncryptCryptoBackend>(
-        self,
-        backend: &mut B,
-        algorithm: iana::Algorithm,
-    ) -> Result<Self, CoseCipherError<B::Error>>;
-}
-
-impl HeaderBuilderExt for HeaderBuilder {
-    fn gen_iv<B: CryptoBackend>(
-        self,
-        backend: &mut B,
-        alg: iana::Algorithm,
-    ) -> Result<Self, CoseCipherError<B::Error>> {
-        let iv_size = aes_algorithm_iv_len(alg)?;
-        let mut iv = vec![0; iv_size];
-        backend.generate_rand(&mut iv)?;
-        Ok(self.iv(iv))
-    }
-}
-
-/// Attempts to perform the COSE operation `op` for a COSE structure with the given `protected` and
-/// `unprotected` headers after verifying that there are no duplicate headers, using the
-/// `key_provider` to determine suitable key candidates.
-///
-/// `key_ops` should be a set of [`KeyOperation`]s that denote that a key is able to perform the
-/// given operation.
-/// Only one of these must be set in the key in order for it to be considered a valid candidate.
-///
-/// If the `key_provider` provides multiple keys, the operation will be attempted for every
-/// candidate until one does not return an error or no more candidates are available.
-///
-/// Errors will be collected into a [`CoseCipherError::NoMatchingKeyFound`], which will be returned
-/// if none of the keys can successfully perform the operation.
-pub(crate) fn try_cose_crypto_operation<BE: Display, CKP: KeyProvider, F, R>(
-    key_provider: &CKP,
-    protected: Option<&Header>,
-    unprotected: Option<&Header>,
-    key_ops: BTreeSet<KeyOperation>,
-    mut op: F,
-) -> Result<R, CoseCipherError<BE>>
-where
-    F: FnMut(
-        &CoseKey,
-        iana::Algorithm,
-        Option<&Header>,
-        Option<&Header>,
-    ) -> Result<R, CoseCipherError<BE>>,
-{
-    if let (Some(protected), Some(unprotected)) = (protected, unprotected) {
-        check_for_duplicate_headers(protected, unprotected)?;
-    }
-    let mut multi_verification_errors = Vec::new();
-    for kc in determine_key_candidates::<CKP, BE>(key_provider, protected, unprotected, key_ops) {
-        multi_verification_errors.push(match kc {
-            Ok((key, alg)) => match op(&key, alg, protected, unprotected) {
-                Err(e) => (key, e),
-                v => return v,
-            },
-            Err(e) => e,
-        });
-    }
-    Err(CoseCipherError::NoMatchingKeyFound(
-        multi_verification_errors,
-    ))
 }
