+++ src/token/cose/util/mod.rs
@@ -1,3 +1,21 @@
-pub(crate) mod header;
-pub(crate) mod ops;
-pub(crate) mod symm;
+/*
+ * Copyright (c) 2024 The NAMIB Project Developers.
+ * Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+ * https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+ * <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+ * option. This file may not be copied, modified, or distributed
+ * except according to those terms.
+ *
+ * SPDX-License-Identifier: MIT OR Apache-2.0
+ */
+mod header;
+mod mac;
+mod ops;
+mod sign;
+mod symm;
+
+pub(crate) use header::*;
+pub(crate) use mac::*;
+pub(crate) use ops::*;
+pub(crate) use sign::*;
+pub use symm::*;
